+++ naga/src/front/wgsl/parse/directive/enable_extension.rs
@@ -6,24 +6,30 @@ use crate::{front::wgsl::error::Error, Span};
 
 /// Tracks the status of every enable-extension known to Naga.
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub struct EnableExtensions {}
+pub struct EnableExtensions {
+    dual_source_blending: bool,
+}
 
 impl EnableExtensions {
     pub(crate) const fn empty() -> Self {
-        Self {}
+        Self {
+            dual_source_blending: false,
+        }
     }
 
     /// Add an enable-extension to the set requested by a module.
-    #[allow(unreachable_code)]
     pub(crate) fn add(&mut self, ext: ImplementedEnableExtension) {
-        let _field: &mut bool = match ext {};
-        *_field = true;
+        let field: &mut bool = match ext {
+            ImplementedEnableExtension::DualSourceBlending => &mut self.dual_source_blending,
+        };
+        *field = true;
     }
 
     /// Query whether an enable-extension tracked here has been requested.
-    #[allow(unused)]
     pub(crate) const fn contains(&self, ext: ImplementedEnableExtension) -> bool {
-        match ext {}
+        match ext {
+            ImplementedEnableExtension::DualSourceBlending => self.dual_source_blending,
+        }
     }
 }
 
@@ -62,7 +68,7 @@ impl EnableExtension {
                 Self::Unimplemented(UnimplementedEnableExtension::ClipDistances)
             }
             Self::DUAL_SOURCE_BLENDING => {
-                Self::Unimplemented(UnimplementedEnableExtension::DualSourceBlending)
+                Self::Implemented(ImplementedEnableExtension::DualSourceBlending)
             }
             _ => return Err(Error::UnknownEnableExtension(span, word)),
         })
@@ -71,11 +77,13 @@ impl EnableExtension {
     /// Maps this [`EnableExtension`] into the sentinel word associated with it in WGSL.
     pub const fn to_ident(self) -> &'static str {
         match self {
-            Self::Implemented(kind) => match kind {},
+            Self::Implemented(kind) => match kind {
+                ImplementedEnableExtension::DualSourceBlending => Self::DUAL_SOURCE_BLENDING,
+            },
+
             Self::Unimplemented(kind) => match kind {
                 UnimplementedEnableExtension::F16 => Self::F16,
                 UnimplementedEnableExtension::ClipDistances => Self::CLIP_DISTANCES,
-                UnimplementedEnableExtension::DualSourceBlending => Self::DUAL_SOURCE_BLENDING,
             },
         }
     }
@@ -83,7 +91,14 @@ impl EnableExtension {
 
 /// A variant of [`EnableExtension::Implemented`].
 #[derive(Clone, Copy, Debug, Hash, Eq, PartialEq)]
-pub enum ImplementedEnableExtension {}
+pub enum ImplementedEnableExtension {
+    /// Enables the `blend_src` attribute in WGSL.
+    ///
+    /// In the WGSL standard, this corresponds to [`enable dual_source_blending;`].
+    ///
+    /// [`enable dual_source_blending;`]: https://www.w3.org/TR/WGSL/#extension-dual_source_blending
+    DualSourceBlending,
+}
 
 /// A variant of [`EnableExtension::Unimplemented`].
 #[derive(Clone, Copy, Debug, Hash, Eq, PartialEq)]
@@ -100,12 +115,6 @@ pub enum UnimplementedEnableExtension {
     ///
     /// [`enable clip_distances;`]: https://www.w3.org/TR/WGSL/#extension-clip_distances
     ClipDistances,
-    /// Enables the `blend_src` attribute in WGSL.
-    ///
-    /// In the WGSL standard, this corresponds to [`enable dual_source_blending;`].
-    ///
-    /// [`enable dual_source_blending;`]: https://www.w3.org/TR/WGSL/#extension-dual_source_blending
-    DualSourceBlending,
 }
 
 impl UnimplementedEnableExtension {
@@ -113,7 +122,6 @@ impl UnimplementedEnableExtension {
         match self {
             Self::F16 => 4384,
             Self::ClipDistances => 6236,
-            Self::DualSourceBlending => 6402,
         }
     }
 }